Clarify state machine of read-csv
This includes several clarifications to the csv parser.
* end-of-field and end-of-line handling is inlined, meaning there
  doesn't need to be a separate variable that tracks those states. The
  loop is a state machine itself, so having a separate state machine was
  extraneous.
* Quoted fields are parsed separately, so that logic is no longer
  interspersed among the other logic.
* Repeated tasks have been extracted into helper functions.

Together these changes result in shorter code that passes the existing
tests. It also fixes some bugs with the handling of headers, quotes,
separators (especially at the beginning and end of a line) and line
endings, (especially at the beginning and end of a file).

It also results, to my eye, in clearer code that is easier to debug.